REF: ExtensionIndex.searchsorted -> IndexOpsMixin.searchsorted
